chore(Team): rename `TeamCard` to `MemberCard`
